rename to contract boxes
