+++ contracts/vending-minter/src/testing/setup.rs
@@ -1,6 +1,6 @@
+pub mod contract_boxes;
 pub mod helpers;
 pub mod msg;
 pub mod setup_accounts_and_block;
 pub mod setup_collection_whitelist;
-pub mod setup_contracts;
 pub mod setup_minter;
