+++ contracts/vending-minter/src/testing/setup/contract_boxes.rs
