+++ contracts/vending-minter/src/testing/tests/whitelist.rs
@@ -13,8 +13,8 @@ use crate::testing::setup::setup_minter::{
     configure_minter, minter_params_all, minter_params_token,
 };
 use crate::testing::setup::{
+    contract_boxes::custom_mock_app,
     setup_accounts_and_block::{setup_accounts, setup_block_time},
-    setup_contracts::custom_mock_app,
 };
 use crate::ContractError;
